server/requirements.txt
absl-py==0.8.0
aniso8601==2.0.1
appdirs==1.4.3
astor==0.8.0
astroid==1.6.1
attrs==17.4.0
backports.functools-lru-cache==1.5
blessings==1.7
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
click==6.7
cloudstorage==0.10.0
configparser==3.5.0
coverage==4.5.1
cycler==0.10.0
enum34==1.1.6
factory-boy==2.10.0
Faker==0.8.11
ffmpeg-python==0.2.0
flake8==3.5.0
Flask==1.0
Flask-Cors==3.0.8
flask-restplus==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.3.2
future==0.17.1
gast==0.3.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-storage==1.19.0
google-pasta==0.1.7
google-resumable-media==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
inflection==0.3.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10.1
jsonschema==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mock==3.0.5
mutagen==1.42.0
nose==1.3.7
numpy==1.16.2
packaging==16.8
pafy==0.5.4
pandas==0.24.1
pluggy==0.6.0
plumbum==1.6.7
protobuf==3.7.1
py==1.5.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.3.1
pycparser==2.19
pydub==0.23.1
pyflakes==1.6.0
pylint==1.8.2
pyparsing==2.3.1
pyrsistent==0.15.4
python-dateutil==2.7.5
python-magic==0.4.15
pytube==9.5.2
pytz==2018.3
requests==2.22.0
rsa==4.0
six==1.12.0
SoundFile==0.10.2
SQLAlchemy==1.3.0
subprocess32==3.5.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
text-unidecode==1.2
urllib3==1.25.3
webapp2==2.5.2
Werkzeug==0.15.3
wrapt==1.10.11
youtube-dl==2019.9.12.1
